refactor: convert credentials.py and scriptgen.py to packages
- credentials.py → credentials/__init__.py
- scriptgen.py → sh/__init__.py (renamed to match CLI command)
- Fix relative imports (single-dot → double-dot for parent package)
- Update all consumers
