+++ tests/e2e_docker/test_pipeline_sh.py
@@ -15,7 +15,7 @@ from nbkp.config import (
 )
 from nbkp.remote.resolution import resolve_all_endpoints
 from nbkp.orchestration import managed_mount
-from nbkp.scriptgen import ScriptOptions, generate_script
+from nbkp.sh import ScriptOptions, generate_script
 
 from nbkp.sync.testkit.seed import build_chain_config
